Add configuration file with path needed for testing.
Problem:
Our tests use pytest-sphinx to create a temporary Sphinx application.
This temporary application needs a source directory that contains a
conf.py file to configure it. This is created in folder test-root
However, pytest-sphinx have not been told to use this folder.

Fix:
Created pytest.ini in root directory, which instructs pytest-sphinx
to use test-root for all Sphinx-related testing.