Fix: Keyboard Navigation & Accessibility Polish
Keyboard navigation with tab and arrows was not working.
The fix centers around usage of radio buttons, that are visually hidden
but must remain interactive. This is achieved with CSS pattern that
makes it invisible, but keeps it in the accessibility tree.
The usage of display:none should not be used for this purpose.

Added styling to highlight selected tab item.